requirements.txt
numpy==1.21.2
Pillow==8.3.1
scour==0.38.2
six==1.16.0
svgwrite==1.4.1
